refactor: extension loading and launching
